+++ mojibake/emoji-sequences.txt
